lib/core/requirements.txt
Django==2.0
django-filter==2.0.0.dev1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
mixer==6.0.0
mock==2.0.0
py==1.4.34
flake8==3.5.0
isort==4.3.4
pyflakes==1.6.0
PyJWT==1.5.3
pytest==3.1.3
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-freezegun==0.2.0
coverage==4.4.1
Pillow==5.0.0
django-dirtyfields==1.3.1
